src/Lwt/JapaneseSegmenter/Natto/requirements.txt
MarkupSafe==2.0.1
Flask==2.0.2
natto-py==1.0.0
